requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.21.0
google-genai>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
